feat: tools moved from example to @fabrice-ai/tools package, fixed header in the usage for medical_survey_server
