+++ example/src/library_photo_to_website.ts
@@ -1,11 +1,10 @@
+import { createFileSystemTools } from '@fabrice-ai/tools/filesystem'
+import { visionTool } from '@fabrice-ai/tools/vision'
 import { agent } from 'fabrice-ai/agent'
 import { teamwork } from 'fabrice-ai/teamwork'
 import { logger } from 'fabrice-ai/telemetry'
 import { solution, workflow } from 'fabrice-ai/workflow'
-import { visionTool } from '@fabrice-ai/tools/vision'
 import path from 'path'
-
-import { createFileSystemTools } from './tools/filesystem.js'
 
 const workingDir = path.resolve(__dirname, '../assets/')
