+++ packages/tools/src/vector/pinecone.ts
